Rename sky_packager to sky_snapshot
This program just creates a snapshot, not the whole package.
