--- sky/tools/packager/switches.cc
@@ -1,13 +1,0 @@
-// Copyright 2015 The Chromium Authors. All rights reserved.
-// Use of this source code is governed by a BSD-style license that can be
-// found in the LICENSE file.
-
-#include "sky/tools/packager/switches.h"
-
-namespace switches {
-
-const char kHelp[] = "help";
-const char kPackageRoot[] = "package-root";
-const char kSnapshot[] = "snapshot";
-
-}  // namespace switches
