+++ sky/tools/sky_snapshot/BUILD.gn
@@ -2,9 +2,7 @@
 # Use of this source code is governed by a BSD-style license that can be
 # found in the LICENSE file.
 
-executable("packager") {
-  output_name = "sky_packager"
-
+executable("sky_snapshot") {
   sources = [
     "loader.cc",
     "loader.h",
