+++ sky/tools/sky_snapshot/switches.h
@@ -1,0 +1,16 @@
+// Copyright 2015 The Chromium Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
+#ifndef SKY_TOOLS_SKY_SNAPSHOT_SWITCHES_H_
+#define SKY_TOOLS_SKY_SNAPSHOT_SWITCHES_H_
+
+namespace switches {
+
+extern const char kHelp[];
+extern const char kPackageRoot[];
+extern const char kSnapshot[];
+
+}  // namespace switches
+
+#endif  // SKY_TOOLS_SKY_SNAPSHOT_SWITCHES_H_
