+++ sky/tools/sky_snapshot/vm.h
@@ -2,12 +2,12 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-#ifndef SKY_TOOLS_PACKAGER_VM_H_
-#define SKY_TOOLS_PACKAGER_VM_H_
+#ifndef SKY_TOOLS_SKY_SNAPSHOT_VM_H_
+#define SKY_TOOLS_SKY_SNAPSHOT_VM_H_
 
 #include "dart/runtime/include/dart_api.h"
 
 void InitDartVM();
 Dart_Isolate CreateDartIsolate();
 
-#endif  // SKY_TOOLS_PACKAGER_VM_H_
+#endif  // SKY_TOOLS_SKY_SNAPSHOT_VM_H_
